backend/requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlmodel>=0.0.16  # Compatible with Pydantic 2.x
psycopg[binary]>=3.2.0  # PostgreSQL driver (works with Neon, Windows-compatible)

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Environment & Config
python-dotenv==1.0.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# AI/ML (Phase 3 - AI Chatbot)
openai>=1.12.0
tiktoken>=0.5.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Development
ipython==8.20.0
